refactor: give comments its own feature package
The comment code was spread across the shared services package, mixed
in with things it has nothing to do with. The other areas already have
their own package, so this brings comments in line and keeps everything
about comments in one place.

It brings its own wiring too, so the startup code just calls into the
package instead of naming every comment class. The reset service now
takes the comments service as an argument, because a plain import
would close a cycle.
