+++ mpvqc/comments/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: mpvQC developers
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from .wiring import bindings as bindings
+from .wiring import register_qml_types as register_qml_types
