+++ mpvqc/comments/models/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: mpvQC developers
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from .store import CommentStore as CommentStore
