+++ mpvqc/comments/services/__init__.py
@@ -2,13 +2,19 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from .roles import ROLE_NAMES as ROLE_NAMES
+from .roles import Role as Role
+from .search import CommentSearchEngine as CommentSearchEngine
 from .search import Found as Found
 from .search import NoMatches as NoMatches
 from .search import NoQuery as NoQuery
 from .search import SearchOutcome as SearchOutcome
 from .selection import SelectionState as SelectionState
 from .service import CommentsService as CommentsService
-from .store import CommentStore as CommentStore
+from .store import StoreItem as StoreItem
+from .time_format_policy import TimeFormatPolicyService as TimeFormatPolicyService
+from .type_validator import CommentTypeValidatorService as CommentTypeValidatorService
+from .types_policy import CommentTypesPolicyService as CommentTypesPolicyService
 from .view_action import AnimatedSelection as AnimatedSelection
 from .view_action import NoViewAction as NoViewAction
 from .view_action import QuickSelection as QuickSelection
