+++ mpvqc/comments/services/store.py
@@ -1,0 +1,42 @@
+# SPDX-FileCopyrightText: mpvQC developers
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Protocol
+
+if TYPE_CHECKING:
+    from collections.abc import Sequence
+
+    from mpvqc.shared import Comment
+
+    from .roles import Role
+
+
+@dataclass(frozen=True, slots=True)
+class StoreItem:
+    comment: Comment
+    seq: int
+
+    @property
+    def sort_key(self) -> tuple[int, int]:
+        return self.comment.time, self.seq
+
+
+class Store(Protocol):
+    def rowCount(self) -> int: ...
+    def item(self, row: int) -> StoreItem: ...
+    def comments(self) -> tuple[Comment, ...]: ...
+    def distinct_comment_types(self) -> frozenset[str]: ...
+    def snapshot(self) -> tuple[StoreItem, ...]: ...
+    def search_rows(self, query: str) -> list[int]: ...
+    def mint(self, comment: Comment) -> StoreItem: ...
+    def insert_position_for_new(self, time: int) -> int: ...
+    def insert_position_for_retimed(self, excluded_row: int, new_time: int) -> int: ...
+    def insert(self, row: int, item: StoreItem) -> None: ...
+    def remove(self, row: int) -> None: ...
+    def replace(self, row: int, new_comment: Comment, role: Role) -> None: ...
+    def move_replace(self, src: int, dst: int, new_comment: Comment, role: Role) -> None: ...
+    def reset(self, items: Sequence[StoreItem]) -> None: ...
