+++ mpvqc/injections.py
@@ -6,13 +6,14 @@ import inject
 
 import mpvqc.services as s
 from mpvqc.appearance import bindings as appearance_bindings
+from mpvqc.comments import bindings as comments_bindings
+from mpvqc.comments.services import CommentsService
 from mpvqc.exporting import bindings as exporting_bindings
 from mpvqc.importing import bindings as importing_bindings
-from mpvqc.services.comments import CommentStore
 
 
-def _comments_service() -> s.CommentsService:
-    return s.CommentsService(inject.instance(CommentStore))
+def _reset_service() -> s.ResetService:
+    return s.ResetService(inject.instance(CommentsService))
 
 
 def _settings_service() -> s.SettingsService:
@@ -21,15 +22,12 @@ def _settings_service() -> s.SettingsService:
 
 def bindings(binder: inject.Binder) -> None:
     appearance_bindings(binder)
+    comments_bindings(binder)
     exporting_bindings(binder)
     importing_bindings(binder)
 
     binder.bind_to_constructor(s.ApplicationPathsService, s.ApplicationPathsService)
     binder.bind_to_constructor(s.BuildInfoService, s.BuildInfoService)
-    binder.bind_to_constructor(CommentStore, CommentStore)
-    binder.bind_to_constructor(s.CommentsService, _comments_service)
-    binder.bind_to_constructor(s.CommentTypesPolicyService, s.CommentTypesPolicyService)
-    binder.bind_to_constructor(s.CommentTypeValidatorService, s.CommentTypeValidatorService)
     binder.bind_to_constructor(s.DesktopService, s.DesktopService)
     binder.bind_to_constructor(s.FileStartupService, s.FileStartupService)
     binder.bind_to_constructor(s.FontLoaderService, s.FontLoaderService)
@@ -40,12 +38,11 @@ def bindings(binder: inject.Binder) -> None:
     binder.bind_to_constructor(s.PlatformService, s.PlatformService)
     binder.bind_to_constructor(s.PlayerService, s.PlayerService)
     binder.bind_to_constructor(s.QuitService, s.QuitService)
-    binder.bind_to_constructor(s.ResetService, s.ResetService)
+    binder.bind_to_constructor(s.ResetService, _reset_service)
     binder.bind_to_constructor(s.ResourceService, s.ResourceService)
     binder.bind_to_constructor(s.SettingsFileService, s.SettingsFileService)
     binder.bind_to_constructor(s.SettingsService, _settings_service)
     binder.bind_to_constructor(s.StateService, s.StateService)
-    binder.bind_to_constructor(s.TimeFormatPolicyService, s.TimeFormatPolicyService)
     binder.bind_to_constructor(s.TimeFormatterService, s.TimeFormatterService)
     binder.bind_to_constructor(s.VersionCheckerService, s.VersionCheckerService)
     binder.bind_to_constructor(s.VideoResizeService, s.VideoResizeService)
