+++ mpvqc/services/__init__.py
@@ -4,9 +4,6 @@
 
 from .application_paths import ApplicationPathsService as ApplicationPathsService
 from .build_info import BuildInfoService as BuildInfoService
-from .comment_type_validator import CommentTypeValidatorService as CommentTypeValidatorService
-from .comment_types_policy import CommentTypesPolicyService as CommentTypesPolicyService
-from .comments import CommentsService as CommentsService
 from .desktop import DesktopService as DesktopService
 from .file_startup import FileStartupService as FileStartupService
 from .font_loader import FontLoaderService as FontLoaderService
@@ -25,7 +22,6 @@ from .reverse_translator import ReverseTranslatorService as ReverseTranslatorSer
 from .settings import SettingsService as SettingsService
 from .settings_file import SettingsFileService as SettingsFileService
 from .state import StateService as StateService
-from .time_format_policy import TimeFormatPolicyService as TimeFormatPolicyService
 from .version_checker import VersionCheckerService as VersionCheckerService
 from .video_resize import VideoResizeService as VideoResizeService
 from .video_resize import ViewDimensions as ViewDimensions
