+++ mpvqc/startup.py
@@ -88,9 +88,10 @@ def register_qml_types() -> None:
     import mpvqc.models  # ruff: ignore[unused-import]
     import mpvqc.viewmodels  # ruff: ignore[unused-import]
     import mpvqc.views  # ruff: ignore[unused-import]
-    from mpvqc import appearance, exporting, importing
+    from mpvqc import appearance, comments, exporting, importing
 
     appearance.register_qml_types()
+    comments.register_qml_types()
     exporting.register_qml_types()
     importing.register_qml_types()
