+++ pyproject.toml
@@ -141,6 +141,22 @@ files = [
     "mpvqc/application.py",
     "mpvqc/build.py",
     "mpvqc/close_event_filter.py",
+    "mpvqc/comments/__init__.py",
+    "mpvqc/comments/models/__init__.py",
+    "mpvqc/comments/models/store.py",
+    "mpvqc/comments/services/__init__.py",
+    "mpvqc/comments/services/commands.py",
+    "mpvqc/comments/services/history.py",
+    "mpvqc/comments/services/roles.py",
+    "mpvqc/comments/services/search.py",
+    "mpvqc/comments/services/selection.py",
+    "mpvqc/comments/services/service.py",
+    "mpvqc/comments/services/store.py",
+    "mpvqc/comments/services/time_format_policy.py",
+    "mpvqc/comments/services/type_validator.py",
+    "mpvqc/comments/services/types_policy.py",
+    "mpvqc/comments/services/view_action.py",
+    "mpvqc/comments/wiring.py",
     "mpvqc/dialogs/__init__.py",
     "mpvqc/dialogs/about.py",
     "mpvqc/dialogs/comment_types.py",
@@ -212,17 +228,6 @@ files = [
     "mpvqc/services/__init__.py",
     "mpvqc/services/application_paths.py",
     "mpvqc/services/build_info.py",
-    "mpvqc/services/comment_type_validator.py",
-    "mpvqc/services/comment_types_policy.py",
-    "mpvqc/services/comments/__init__.py",
-    "mpvqc/services/comments/commands.py",
-    "mpvqc/services/comments/history.py",
-    "mpvqc/services/comments/roles.py",
-    "mpvqc/services/comments/search.py",
-    "mpvqc/services/comments/selection.py",
-    "mpvqc/services/comments/service.py",
-    "mpvqc/services/comments/store.py",
-    "mpvqc/services/comments/view_action.py",
     "mpvqc/services/desktop.py",
     "mpvqc/services/file_startup.py",
     "mpvqc/services/font_loader.py",
@@ -266,7 +271,6 @@ files = [
     "mpvqc/services/settings.py",
     "mpvqc/services/settings_file.py",
     "mpvqc/services/state.py",
-    "mpvqc/services/time_format_policy.py",
     "mpvqc/services/version_checker.py",
     "mpvqc/services/video_resize.py",
     "mpvqc/shared.py",
