+++ test/comments/services/test_add.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from mpvqc.services.comments import NoViewAction, QuickSelection, QuickSelectionAndEdit
+from mpvqc.comments.services import NoViewAction, QuickSelection, QuickSelectionAndEdit
 
 
 def test_add_comment(comments):
