+++ test/importing/services/test_importer.py
@@ -10,6 +10,7 @@ from unittest.mock import MagicMock
 import inject
 import pytest
 
+from mpvqc.comments.services import CommentsService
 from mpvqc.importing.services import (
     ImportService,
     ImportSettingsService,
@@ -23,7 +24,7 @@ from mpvqc.importing.services import (
     VideoLoad,
     VideoSkip,
 )
-from mpvqc.services import CommentsService, PlayerService, ResetService, StateService
+from mpvqc.services import PlayerService, ResetService, StateService
 from mpvqc.shared import Comment
 from test.importing.plans import (
     SUB_A,
