+++ test/test_import_rules.py
@@ -14,7 +14,7 @@ import pytest
 
 REPO = Path(__file__).resolve().parents[1]
 
-SLICES = ("appearance", "exporting", "importing")
+SLICES = ("appearance", "comments", "exporting", "importing")
 COMPOSITION_ROOTS = "mpvqc/injections.py and mpvqc/startup.py"
 HELPERS = ("jobs",)
 SHARED_ROLES = {"services": "services", "shared": "shared"}
